+++ pwd.php
@@ -3,18 +3,12 @@ include('includes/db.php');
 
 session_start();
 $id = $_SESSION['id'];
+$username = $_SESSION['username'];
+$barangay_id = $_SESSION['barangay_id'];
 
 if (!isset($_SESSION['id'])) {
     header('location: login.php');
 }
-
-// Fetch Barangay ID from Users
-$get_data = "SELECT * FROM users WHERE id = $id";
-$run_data = mysqli_query($conn, $get_data);
-$fetch = mysqli_fetch_assoc($run_data);
-
-$barangay_id = $fetch['barangay_id'];
-// Fetch End
 
 // Get barangay id from barangays
 $sql = "SELECT * FROM barangays WHERE barangay_id = '$barangay_id'";
@@ -30,16 +24,34 @@ $barangay_name = $row['barangay_name'];
 if (isset($_GET['delete_resident'])) {
     $id = $_GET['delete_resident'];
     $purok_id = $_GET['purok_id'];
+    $username = $_GET['username'];
 
-    $sql = "DELETE FROM residents WHERE id=$id";
-
-    $result = mysqli_query($conn, $sql);
-
-    if ($result) {
-        header('location: pwd.php?delete=success');
-        exit;
+    if (empty(trim($_POST['password']))) {
+        $password_err = "Please enter your password!";
     } else {
-        header('location: pwd.php?delete=failed');
+        $password = mysqli_real_escape_string($conn, $_POST['password']);
+    }
+
+    if (empty($password_err)) {
+        $password = md5($password);
+
+        $sql = "SELECT * FROM users WHERE username = '$username' AND password = '$password'";
+        $result_query = mysqli_query($conn, $sql);
+
+        if (mysqli_num_rows($result_query) == 1) {
+            $sql = "DELETE FROM residents WHERE id=$id";
+
+            $result = mysqli_query($conn, $sql);
+
+            if ($result) {
+                header('location: pwd.php?delete=success');
+                exit;
+            } else {
+                header('location: pwd.php?delete=failed');
+            }
+        } else {
+            $password_err = "Incorrect Password!";
+        }
     }
 }
 
@@ -156,6 +168,8 @@ if (isset($_POST['edit_resident'])) {
         <div class="alert alert-success text-center m-3 edit"><span>Data has been Edited!</span></div>
     </div>
 
+    <?php include('includes/delete_error.php'); ?>
+
     <div class="dashboard-content pt-4">
         <div class="table-container table-responsive p-2">
             <table class="table table-striped table-hover table-sm text-center align-middle" id="datatable">
@@ -230,32 +244,32 @@ if (isset($_POST['edit_resident'])) {
 
     <!-- Edit Modal -->
     <?php
-        $sql = "SELECT * FROM residents";
-        $result = mysqli_query($conn, $sql);
-        while ($row = mysqli_fetch_array($result)) {
-            $resident_id = $row['id'];
-            $first_name = $row['first_name'];
-            $middle_name = $row['middle_name'];
-            $last_name = $row['last_name'];
-            $residents_address = $row['residents_address'];
-            $gender = $row['gender'];
-            $dob = $row['dob'];
-            $civil_status = $row['civil_status'];
-            $occupation = $row['occupation'];
-            $school_attainment = $row['school_attainment'];
-            $skills = $row['skills'];
-            $blood_type = $row['blood_type'];
-            $citizenship = $row['citizenship'];
-            $purok_id = $row['purok_id'];
-            $household_type = $row['household_type'];
-            $four_ps = $row['4p_s'];
-            $pwd = $row['pwd'];
+    $sql = "SELECT * FROM residents";
+    $result = mysqli_query($conn, $sql);
+    while ($row = mysqli_fetch_array($result)) {
+        $resident_id = $row['id'];
+        $first_name = $row['first_name'];
+        $middle_name = $row['middle_name'];
+        $last_name = $row['last_name'];
+        $residents_address = $row['residents_address'];
+        $gender = $row['gender'];
+        $dob = $row['dob'];
+        $civil_status = $row['civil_status'];
+        $occupation = $row['occupation'];
+        $school_attainment = $row['school_attainment'];
+        $skills = $row['skills'];
+        $blood_type = $row['blood_type'];
+        $citizenship = $row['citizenship'];
+        $purok_id = $row['purok_id'];
+        $household_type = $row['household_type'];
+        $four_ps = $row['4p_s'];
+        $pwd = $row['pwd'];
 
-            if ($household_type === "Head") {
-                $household_type_result = "Head of Household";
-            } else {
-                $household_type_result = "Member of Household";
-            }
+        if ($household_type === "Head") {
+            $household_type_result = "Head of Household";
+        } else {
+            $household_type_result = "Member of Household";
+        }
 
         echo "
                 <div class='modal fade' id='edit$resident_id' data-bs-backdrop='static' data-bs-keyboard='false' tabindex='-1' aria-labelledby='staticBackdropLabel' aria-hidden='true'>
@@ -387,58 +401,32 @@ if (isset($_POST['edit_resident'])) {
     }
     ?>
 
-    <!-- Delete Resident Modal -->
+    <?php include('includes/delete_modals.php'); ?>
+
+
+    <!-- Residents Profile Modal -->
     <?php
     $sql = "SELECT * FROM residents";
     $result = mysqli_query($conn, $sql);
     while ($row = mysqli_fetch_array($result)) {
         $resident_id = $row['id'];
-
-        echo "
-                <div class='modal fade' id='delete$resident_id' data-bs-backdrop='static' data-bs-keyboard='false' tabindex='-1' aria-labelledby='staticBackdropLabel' aria-hidden='true'>
-                    <div class='modal-dialog'>
-                        <div class='modal-content'>
-                            <div class='modal-header bg-dark text-light'>
-                                <h5 class='modal-title' id='staticBackdropLabel'>Are you sure you want to delete?</h5>
-                                <button type='button' class='btn-close bg-light' data-bs-dismiss='modal' aria-label='Close'></button>
-                            </div>
-                            <div class='modal-body'>
-                                <h6>Resident data will be deleted.</h6>
-                            </div>
-                            <div class='modal-footer'>
-                                <a href='pwd.php?delete_resident=$resident_id' type='button' class='btn btn-danger'>Delete</a>
-                                <button type='button' class='btn btn-secondary' data-bs-dismiss='modal'>Cancel</button>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            ";
-    }
+        $first_name = $row['first_name'];
+        $middle_name = $row['middle_name'];
+        $last_name = $row['last_name'];
+        $residents_address = $row['residents_address'];
+        $gender = $row['gender'];
+        $dob = $row['dob'];
+        $citizenship = $row['citizenship'];
+        $civil_status = $row['civil_status'];
+        $occupation = $row['occupation'];
+        $school_attainment = $row['school_attainment'];
+        $skills = $row['skills'];
+        $blood_type = $row['blood_type'];
+        $purok_id = $row['purok_id'];
+        $household_type = $row['household_type'];
+        $four_ps = $row['4p_s'];
+        $pwd = $row['pwd'];
     ?>
-
-    <!-- Residents Profile Modal -->
-    <?php
-        $sql = "SELECT * FROM residents";
-        $result = mysqli_query($conn, $sql);
-        while ($row = mysqli_fetch_array($result)) {
-            $resident_id = $row['id'];
-            $first_name = $row['first_name'];
-            $middle_name = $row['middle_name'];
-            $last_name = $row['last_name'];
-            $residents_address = $row['residents_address'];
-            $gender = $row['gender'];
-            $dob = $row['dob'];
-            $citizenship = $row['citizenship'];
-            $civil_status = $row['civil_status'];
-            $occupation = $row['occupation'];
-            $school_attainment = $row['school_attainment'];
-            $skills = $row['skills'];
-            $blood_type = $row['blood_type'];
-            $purok_id = $row['purok_id'];
-            $household_type = $row['household_type'];
-            $four_ps = $row['4p_s'];
-            $pwd = $row['pwd'];
-        ?>
         <div class="modal fade" id="view<?php echo $resident_id; ?>" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
             <div class="modal-dialog">
                 <div class="modal-content">
@@ -528,12 +516,12 @@ if (isset($_POST['edit_resident'])) {
 
     <!-- Print Modal -->
     <?php
-        $sql = "SELECT * FROM residents";
-        $result = mysqli_query($conn, $sql);
-        while ($row = mysqli_fetch_array($result)) {
-            $resident_id = $row['id'];
+    $sql = "SELECT * FROM residents";
+    $result = mysqli_query($conn, $sql);
+    while ($row = mysqli_fetch_array($result)) {
+        $resident_id = $row['id'];
 
-        ?>
+    ?>
         <div class="modal fade" id="print<?php echo $resident_id; ?>" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
             <div class="modal-dialog">
                 <div class="modal-content">
@@ -543,7 +531,7 @@ if (isset($_POST['edit_resident'])) {
                     </div>
                     <div class="cert-modal-body modal-body">
                         <ol>
-                        <li>
+                            <li>
                                 <a href="certs/barangay_clearance.php?resident_id=<?php echo $resident_id; ?>&&purok_id=<?php echo $purok_id; ?>&&pwd">Barangay Clearance</a>
                             </li>
                             <li>
